+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 ## Changes 4//13/2025
 
 - Decreased header height some more and clickable logo.
+- authModals.js fully updated with i18n.js keys.
+- main.js added Dark & Light mode i18n.js keys.
 
 ## Changes 4/12/2025